pdf converter / editable (NOT FINISHED)
git-svn-id: http://www.pimcore.org/svn/pimcore/private/core/trunk@2809 1f8fe7d8-47f0-464c-8d0a-336f4953ab05

+++ pimcore/models/Asset/Image.php
@@ -69,11 +69,14 @@ class Asset_Image extends Asset {
     /**
      * @return void
      */
-    public function clearThumbnails() {
-        $files = glob(PIMCORE_TEMPORARY_DIRECTORY . "/thumb_" . $this->getId() . "__*");
-        if(is_array($files)) {
-            foreach ($files as $file) {
-                unlink($file);
+    public function clearThumbnails($force = false) {
+
+        if($this->_dataChanged || $force) {
+            $files = glob(PIMCORE_TEMPORARY_DIRECTORY . "/thumb_" . $this->getId() . "__*");
+            if(is_array($files)) {
+                foreach ($files as $file) {
+                    unlink($file);
+                }
             }
         }
     }
